rename r1cs_reader to r1cs_parser
