+++ src/lib.rs
@@ -12,5 +12,5 @@ extern crate rand;
 pub mod circom_circuit;
 pub mod plonk;
 pub mod plonk_util;
-pub mod r1cs_reader;
+pub mod r1cs_parser;
 pub mod reader;
